Fix find handles to deleted file test for Windows 10 1903 (#152)
When BuildXL fails to delete a file due to access denied issues or similar, it attempts to find all open handles to the file for logging.

Previously, if the file was in the Windows "pending deletion" state (file deleted, but there is an open handle with FileShare.Delete permissions to it somewhere), Windows would thrown an access exception if any attempt to create a new file handle to the file was created. BuildXL's attempts to find open handles would also fail to detect any existing file handles open.

The test FileUtilitiesUnsafeTests.FailToFindAllOpenHandlesPendingDeletion pins this behavior.

With Windows 10, Version 1903, Windows no longer throws an access exception in the above scenario and the File is considered deleted by C# File.Exists as long as all existing handles have FileShare.Delete permissions. but BuildXL is still unable to detect any existing file handles open.

This adds an or condition to fix the assumptions of the test to handle Version 1903.

Fixes AB#1496544