Correctly compute priority by separating critical path priority from a given spec priority (#388)
When priority was first added, the computation was done incorrectly. There are two pieces to priority, the one specified in the spec file and the one computed by critical path. To compute the new values for these, the outgoing pips priority needs to be separated into the specified priority and the critical path priority. Each of these values in then maxed to find the largest priority that the pip should have.