Retry process if exits with Azure Watson 0xDEAD exit code (#401)
Process can nondeterministically exit with 0xDEAD exit code in CB. This is the exit code returned by Azure Watson dump after catching the process crash. The root cause of the crash is unknown, but the primary suspect is the way Detours handle NtClose. One needs to have a deep investigation on detoured NtClose.

This change is an attempt to improve reliability by retrying the process.