Dropped directories may be materialized lazily (#420)
There're 3 lines of code that require directories to drop to exist on disk before the drop method is called. We already have lazily materialized output directories. Thus, the existence requirement is no longer needed. In fact it's causing an error because when the output directories are lazily materialized, then those directories may not be on disk yet.