Order pip fingerprints when adding to json (#446)
Adding order independent items to the fingerprinter should actually sort them before adding them to the json text. To do this, a comparer is given for FileArtifacts, FileArtifactsWithAttributes, and DirectoryArtifacts. The actual pip fingerprinter ignores this comparer as it XORs the values together, but the json fingerprinters sorts the items before adding them.