Changes needed to run tests in VM (#457)
These changes are needed to run tests in VM.

One notable change is the way we check the ability to scan the journal. Normally, we check the root path of the executing assembly to know if journal scanning can be performed. However, in the VM scenario the executing assembly is on the host, and one cannot check journal capability of the host from VM.

Another less safe change is skipping file tracking if the root path is in a journal incapable volume. Our unit tests (particularly the integration tests) track the test executable (the one that execute Operations.cs). But that executable is located on the host, and tracking the host USN from VM does not work. Unfortunately, once tracking failed, the file change tracker switches into disabled state, and this failed all incremental scheduling tests.