Allow for error regex to match across multiple lines (#825)
Enable multi line error scanning. 

When enabled: process output is scanned for error messages line by line; 'errorRegex' is applied to each line and if ANY match is found the ENTIRE line is reported.

When disabled: process output is scanned in chunks of up to 10000 lines; 'errorRegex' is applied to each chunk and only the matches are reported.  Furthermore, if 'errorRegex' contains a capture group named "ErrorMessage", the value of that group is reported; otherwise, the value of the entire match is reported.

Because this scanning is done against chunks of text (instead of the entire process output), false negatives are possible if an error message spans across multiple chunks.  The scanning is done in chunks because attempting to construct a single string from the entire process output can easily lead to OutOfMemory exception.