Change working directory for running VmCommandProxy (#887)
Currently when running VmCommandProxy, the working directory is set to the working directory of the pip. However this can "conflict" with the operation that the pip will do to that working directory.

VmCommandProxy writes log files to that working directory. QTest pips set their working directory to the deployment directory. Later the pips robocopy this directory. The robocopy process will cause DFA when it finds the logs written by VmCommandProxy.