Reduce the time between getting a port for API server and starting the server. (#1251)
When we create an AI server, we render its moniker (i.e., getting an available port from the OS) and pass it as an argument. Usually this works fine; however, if scheduler initialization takes longer, the OS might mark the port as unused, AND if another process asks for a free port, our port (the one that we rendfered previously) might be given to that process. As a result, when we start the server, it will fail because the port is already in use.

AB#1617116