Merged PR 521873: Ensure that different repositories get different app servers
App server carries in-memory state, e.g., incremental scheduling state, that should not be shared between repositories.

This change ensures that different repositories get different app servers. To this end, we include the full path of the config, instead of just a relative path, in the server's name. Moreover, we include subst drive and subst source in the server's name.

Related work items: #1652012