Merged PR 522629: Use cheaper hash algorithm for server deployment
This fixes a server mode crash that happens when the client fails to connect to an existing server. The crash is a race condition between selecting the hash algorithm and the hasher pool size. So not using the standard VSO hasher in ContentHashingUtilities removes that race condition.

This change is effective at addressing the crash but exposes the fragility in our hash algorithm initialization code. I'm going to rework that separately but I want to get the crash addressed.

Related work items: #1653277