Merged PR 523808: Tune ApproximateSort to remove a fraction of the maximum entries in the pool
We have observed in production that the pool can occasionally fill with bad candidates. In such an scenario, there's nothing we can actually do but return them. This adds a knob that lets us remove a fraction of bad candidates from the pool.

The key assumption here is that the fraction of the max candidates we purge from the pool is small enough that it won't cause eviction to fail because of missing candidates.