Merged PR 524609: Cancel Raided Redis Test and Bug Fix
We have a counter to check when we cancel one of the two redis instances.
The test performs a get bulk async first and verifies the counter is zero.
Then we close the secondary redis database, and set a retry timeout window. Because the secondary database is closed, we should result to the timeout window and incrementing the counter value.
Check for a counter value of 1.

Bug Fix: We noticed that certain redis operations were taking longer than the set retry window value in CBTest. This is because when we are getting a task back from ExecuteAndCaptureRedisErrorAsync, we are awaiting the callback of executeAsync, but if the synchronous code in this callback will execute first before we return the task. Therefore if the callback has synchronous code that takes a long time, the task will take a long time before returning, resulting in a long redis operation. To fix this we will yield the task before calling the callback, allowing us to continue with the cancel redis logic, until we await for the callback task.

Also pass severity during fire and forget message in raidedRedisDatabase, and changed tracing to reflect we're cancelling a redis instance.