Merged PR 524711: Introduce external sandboxed process directory to ensure no path collision between tools for external executions and pips
Previously, when executing pips externally (e.g., in VM), the working directory of the tools used for external executions (e.g., `SandboxedProcessExecutor` and `VmCommandProxy`) is set to the pip's std directory, i.e., the directory where the pip will place its std.out. This working directory is used by these tools to place their own logs, stdout, and stderr. Note that these are not the artifacts produced by the pip, but by the tools that enable the pip's external execution.

The problem with sharing the pip's std directory with the tools working directory is the pip's std directory can be specified as an output directory of the pip. The tool that the pip executes can freely delete and recreate that directory, like `QTest`. But, in that scenario, the deletion will fail because the tools that enable the external execution have already produced files there and are potentially still holding handles to those files until the pip finishes its execution. An example of those files is `VmCommandProxy.log`.

This change creates a separate directory for tools that enable external executions to place their own files. This directory is disjoint from where pips produces their output artifacts.

This change is a prerequisite for running QTest in VM.

Example of build where QTest runs in VM: https://b/build/bf791c01-a175-4a18-a090-c82662708c72

Related work items: #1657003