Merged PR 523527: Copy retry bug fix with relation to destination path error
Purpose of work item: Originally we calculated a waitDelay between attempts of copying from a list of locations, and after each attempt we wait for the full length of waitDelay. Now instead we compute a waitDelay for each attempt of copying, but for each location we try to copy from, we keep track of how long the location has already waited while the other locations finish operating, and only wait the remaining time to match the computed waitDelay.

Example Case: Locations A, B, C taking 3 seconds each to operate, and we compute a waitDelay of 7 seconds for next attempt.

Previous method: Wait 9 seconds for A, B, C to complete, then wait another 7 seconds before second attempt at A, so total of 16 seconds before starting second attempt of A.

New method: Wait 9 seconds for A, B, C to complete, but since A already waited 6 seconds for B and C to complete, it only waits (7-6) = 1 second for its second attempt, for a total of 10 seconds before starting second attempt of copying A.

Bug: Happens in the first attempt of copying from a list of locations, and one location results in a destination path error. In following attempts of copying, if that same location now produces a different result, there will be an out of range exception thrown.

- Added test case that runs through this case. Has a destination path error as a result in the first attempt of copying, then in the following attempts of copying should still run properly without out of range.

- DistributedContentCopier is fixed by initializing an array to certain amount of elements before passing into WalkLocationsAndCopyAndPutAsync.