Merged PR 525343: Prepare for the wdg vpack moving from the full framework version of bxl to the
Prepare for the wdg vpack moving from the full framework version of bxl to the dotnet core version.
In the netcore vpack for osgtools we have to ship some of the net472 bits as osgtools uses many assemblies as libraries and they compile for net48.
I created a net472 subfolder so we sets the precedent that when osgtools moves some of their tools to net core, that we can add a netcore sibbiling folder and extend without breaking.
Note our release pipeline that produces the vpack combines the `out\bin\${configuration}\private\wdg` folder with the `out\bin\${configuration}\${targetruntime}`