Merged PR 525636: Add native support for managed shared compilation
This PR enables native support for managed shared compilation (that is, VBCSCompiler) for the bxl self-host. The option is exposed as part of bxl.cmd (off by default for now):

bxl.cmd -UseManagedSharedCompilation

(turning on this flag results in a build that is twice as fast for -minimal :))

This is reusing the existing machinery to configure VBCSCompiler to breakaway from the sandbox, plus a new native way to compensate missing accesses for trusted processes by using the statically declared artifacts of a pip to pretend the (breakaway) process actually made those accesses.

This is for now only exposed for csc at the managed SDK level, but it could be used for other service-like tools (e.g. mspdbsrv, vctip, etc.)