Merged PR 526296: Fix when RocksDb logs backup happens
The current code will backup right before loading, but the argument passed in is the new incarnation of the DB, which will never have any logs. This means the mechanism is effectively not doing anything except for the first iteration, in which fast loading of the previous checkpoint means we will have some logs to backup.