Merged PR 526542: Make RocksDbMemoizationStore log less in dev builds
The default log policy for RocksDb was to keep logs for 30 days, which works fine for servers, but is way too long for dev builds. This lead to the number of logs being stored having a size significantly larger than the actual metadata DB.

This PR makes us keep a single file, up to 1 MB in size, and keeping only the last hour (just in case it might come in handy for debugging at some point).