Merged PR 526964: Scrubbing files for deployments and unittests
Scrubbing files for deployments and unittests.
When deploying we were partially sealing the folders. This meant that files no longer deployed were left on disk. This frequently lead to DFA's in unittests when the test framework insist on relfecting over all the files in the test folder to load items.
This change takes advantage of the earlier shipped feature that allows fully sealed directory to delete files that are on disk but not part of that folder to ensure the folders are pure.
The feature is enabled by default in `Deployment.deployToDisk`. You can use the option `sealPartialWithoutScrubbing` if you needed this.