Merged PR 525799: Content Repudation
Provides the function of distributed delete for a certain content. When one machine receives the delete command, it starts by creating a GrpcContentClient that passes the deleteOption of deleting distributed. We start by calling the local DeleteAsync on the original machine, then we register the delete to event hub, and check for other machines that have the same content. For each machine that has the same content, we create a new GrpcContentClient that calls delete, but this time the deleteOption is locally. Those machines call the local DeleteAsync, and each tell event hub when they're done deleting.

Test:
-  Put the same file on a variable number of machines, one being master rest are workers.
- Heartbeat the master to create checkpoint, and heartbeat workers to restore checkpoints
- Verify correct number of machine locations have this file based off master
- Call delete async on one of the workers, which should perform a distributed delete on every machine with the content
- Verify the content is no longer recorded on any machine from master