Merged PR 526969: Fix incorrect file whitelisting by introducing CreateProcess reported access
If a process X, with executable path `x`, launches another process Y, with executable path `y`, but there is a whitelist `{ tool: y, pattern: * }`, then BuildXL says that the access of Y matches the whitelist. It is because there is only one reported file access related to this activity, which is
```
{
   Operation: Process
   Process: 'Process(Y)'
   Path: 'y'
}
```
With the above reported access, `Process(Y)` has path `y` which matches the whitelist entry.

Instead of setting `Process` to `Process(X)` (the parent) in the above reported access, I decided to introduce another operation, called `CreateProcess`. The reason being is getting parent id is not reliable when the child process is detached.

Unlike Process operation that is sent by the created process, CreateProcess operation is sent by the creating process and thus has a proper process id. With this approach, we can now validate the eligibility of spawning a process, as well as enforcing symlink reachability on the executable.

The down side of this approach is the use of logic for inferring the executable path.

Related work items: #1661067