Merged PR 526720: Add /CacheOnly mode which skips execution upon cache misses
The /cacheonly mode will skip pips that are cache misses instead of executing them. There have been various requests for this feature in different contexts. It is helpful for debugging cacheability of builds and opens up some possible use cases for DevCache.

Apart from the new feature itself, the big change here is that Skipped pips are no longer considered failures. Previously the only cause of a skipped pip was an upstream failed pip but now that's no longer the case. I considered adding an additional state but I didn't want to complicate the user experience and implementation with having a new concept in addition to the pre-existing skipped state.

Related work items: #1591138