Merged PR 526872: Use absolute local time in log files in CloudBuild
Most logs in CloudBuild are in terms of absolute local (redmond) timezone time. The BuildXL log file is an outlier in terms of relative time from the beginning of the build. This makes it inconvenient to correlate BuildXL logs with other logs such as the cache, BuildRunners, CB service logs, etc.

This change only impacts log files. It adds an absolute time prefix in the same log4net format used by most other CloudBuild components. It also retains the relative time in seconds granularity in case people are used to that
```2020-01-03 19:36:18,359 [0:20] verbose DX3627: [Pip9998B10CD07CA998, <COPYFILE```...

The console is unaffected. It remains in relative time.

Related work items: #1323731