Merged PR 527903: Demote pathset parse failure to verbose level
When we change the serialization format of the pathset (like we just did recently) this event is triggered on the build that is unable to use the previous build's metadata record. It is useful to know this happens for debugging but it is expected in scenarios like this and there's no user action/attention necessary, so a warning seems too much. Demote it to a verbose level event.