Merged PR 528237: Keep a list of hierarchical names for the cache id instead of a flattened string
Providing cache id names with underscores (the case for CB, whose short name is defined as 'cb_m') causes a bxl crash. This is because the cache id represents a hierarchical name, but it was stored as a string using underscores to split each component of the id. This PR changes the cache id to be a more structured entity.