Merged PR 529003: Fix leftover of cache id type refactoring
This is a leftover of the cache id type change that was made a couple of days ago. It should only be affecting config cop testing (this is serialization related), which has already been fixed on the CB side.

I'm not aware of any other users of cache id default values.