Merged PR 530602: Cache weak fingerprint JSON serialization
JSON serialization over weak fingerprint can happen twice when there is a cache miss: one for cache-lookup event and one for execution event.

As such a serialization can be very expensive, we cache it in a transient cache so that it doesn't consume a lot of memory.

We also add detailed counters about run-time cache miss analysis.

Related work items: #1475539