Merged PR 530854: Don't include queue time in ProcessDuration counters
I was looking at pip counter telemetry and noticed that the ProcessDurationMs was way higher than ExecuteProcessDurationMs. Looking at the implementation it seems we were including the time the pip was sitting in a queue when BuildXL was't actually working on it. I think this revised definition of the counter is more useful