Merged PR 531032: Support running crossgen on managed binaries
This PR enables the option to run crossgen (aka ReadyToRun https://docs.microsoft.com/en-us/dotnet/core/whats-new/dotnet-core-3-0#readytorun-images) on managed binaries. Not on by default for now. This is only available for dotnetcore apps (target runtime version 3.0 preview 5 and above).

It can be enabled for our selfhost with: bxl.cmd /p:[Sdk.BuildXL]enableCrossgen=1

I'll do some perf analysis and validations before trying to enable this for any real scenario.

Note: for now crossgen does not support cross-targeting, which means that, for example, in order to get ready to run images for mac, they need to be compiled on mac