Merged PR 531023: UseBinManager under flag
UseBinManager under flag.

Added a stress test to validate that if we remove the locks in BinManager, it will not break serialization.

Also, wrap deserialization in a try/catch block to make sure that if it fails, we just continue with the latest bin manager. If this happens to the master and it is its first restore checkpoint since startup, that means it will restart the BinManager from scratch.