Merged PR 531771: Fix crash when there is an IO error processing tool output streams
TryFilterAsync() historically had a return type of string. In error cases it returned null and its callers knew how to handle that. Recently it was changed to return an object and the callers were updated to access a field in that object. The error paths in TryFilterAsync() still returned null resulting in the consumers hitting NREs when those error codepaths were exercised.

Related work items: #1670314