Merged PR 532107: Enable QTest for some more tests
Enable QTest for
 - Test.Tool.MsBuildGraphBuilder.dsc
 - Test.Tool.VBCSCompilerLogger.dsc
 - Test.BuildXL.FrontEnd.Download.dsc
 - Test.BuildXL.Storage.dsc.

No other changes were needed for the first three (not sure why qtest wasn't enabled there in the first place).

For the storage tests, created symlinks must be manually deleted to prevent a subsequent call to `robocopy` by `qtest` from failing (see !531660).  Additionally, when run as admin, the tests leave something around that causes qtest to fail with
```
System.IO.IOException: The data present in the reparse point buffer is invalid.
```
To work around that, introduce a flag forcing the xunit framework to be used for admin tests.

Related work items: #1668899