Merged PR 532149: Enable code coverage for undetoured tests
The big thing holding us back from coverage was that child processes of tests would access QTest files when they were injected with code coverage. This threw off our tests because many children of test processes failed with DFAs.

This change tells vstest to not track coverage for child processes across the board. Before finding this setting I attempted to address this by making our tests mark the QTest code coverage assemblies as untracked. I left this change in on the off chance that we ever do want to gather coverage for child processes of tests. But as things stand now it shouldn't be necessary.

There was one test which code coverage exposed a race condition in. That test wasn't actually testing our code as much as it was pinning our understanding of Windows' behavior. I did a low-effort tweak to make it pass by utilizing the more robust file deletion utility on test cleanup.

Related work items: #1668900