Merged PR 531746: Modify copy loop to test reducing pin threshold
Modify copy loop to test reducing pin threshold.

The intent is to simulate what would happen if we were to reduce the pin threshold. If the copy completes within the restricted loops, it means it is safe to reduce the pin threshold. Otherwise, it would be unsafe to reduce it, since copies actually need the extra replicas

Related work items: #1670195