Merged PR 531483: Track paths used for weak fingerprint augmentation
To create an augmented weak fingerprint, we first need compute an augmented pathset. This pathset is used every time we do a cache lookup to compute an augment weak fingerprint of a pip.

It's possible that a pip might stop consuming a path that was included in the augmented pathset. If such a path is a file that no longer exists, it's not a big problem (just one-time cache miss). However, if this path is changing every build, this would prevent us from having any cache hits against the augmented pathset.

This PR adds an option to monitor the usage of the paths in the augmented pathset, and log any paths that were not observed during pip execution.

Related work items: #1643752