Merged PR 532631: Stop running unittest on Net472 for code that is not consumed as net472.
Stop running unittest on Net472 for code that is not consumed as net472.

BuildXL.Processes.dll is still used by CloudBuild in net472, so I kept that test running on net472. This unfortunatley forced some of the utilities we use as test processes to remain in net472 because they are copied next ot the test dll. so if I moved those to always be netcoreapp3.1 there would be copy violations of the nuget dependencies. Therefore thye remained in net472 as well. The reason is commented so when we stop needing to produce buildxl.process as net472 we can easily remove these.