Merged PR 534057: Add 'Nullable' property for old vs generation if `nullable: true` is provided
Currently, we're using an old spec generator by default, so if a solution is generated by running `bxl /vs` then the projects that specified `nullable: true` won't have `<Nullable>enabled</Nullable>` section in the generated project file.

This PR adds the same nullability support for the old solution generation that was already added to the new solution generation.

P.S. @<Aleksandar Milicevic>  Do we have a timeframe when we'll retire an old generator? Or should we reuse the logic in both of them? Because I wasn't aware of the old one and when I added support for nullable types, I've added it only in the new one, and back then that was the default we used. But the generation was broken once we switched back to the old one by default.