Merged PR 534314: Fix reverse eviction order and make sure we don't accept content which is older than what we just evicted
### Fix reverse eviction order
At some point we stopped respecting the "reverse" parameter when getting content in eviction order. To fix this, create a new instance of the comparer which returns the inverse of what eviction would want.

### Make sure we don't accept content which is older than what we just evicted
Every time the distributed layer unregisters a location, it will remember its last access time, so that when it is attempting to receive a pushed copy, it makes sure that it is not older than what we are evicting.

Some refactors had to be done: instead of having a "HasContentLocally" method, there is a "CanAcceptContent", which returns an enum saying why the content has been rejected from the machine (Including ContentAlreadyLocal to preserve previous behavior), or NotRejected if the store can receive the content.

###Tests
Since some of the logic lives in the gRPC layer, it was necessary to add tests that actually communicated via gRPC. Infrastructure was added to make this possible, by overriding the UseGrpcServer flag and set it to true.