Make File.ReadAllText synchroneous since in office build we run out of threads
Make File.ReadAllText synchroneous since in office build we run out of threads to handle the async IO callback.

The real fix is to use the ActionBlock threading model for the frontend so we have fair scheduling... I have started some flights to test the perf impact of makign that the default.