Merged PR 536368: Scrubber can follow directory symlinks
This PR extends the scrubber behavior to include directory symlinks. This is useful for the case of shared opaque directories containing directory symlinks. In the past, the scrubber wouldn't traverse these. Additionally, on Windows, the scrubber was deleting all symlink directories under shared opaques since it was considering them empty (since they were not explored).