Merged PR 535265: [WIP] Flaky tests validation
Using PR Validation to check tests for flakiness, will push multiple changes to run the test multiple times. Tests that are changed and fixed due to flakiness and race conditions.  DualRedundancyGlobalRedisTest and ProactiveReplicationTest.

DualRedundancyGlobalRedisTest had a possible race condition in the test just at startup before we even mirror the cluster states for master. This ends up resulting in mismatch for machine location between machine and worker.

ProactiveReplicationTest simulates performing proactive replications in the startup of the sessions. We only wanted to test replicating once, but with extra machines we can replicate multiple times occasionally.