Add kill switch for ContentLocationDatabase's replace
The issue being tackled is:
- Long-running operations hold onto a RW lock in read mode.
- When a replacement of the database is in order (i.e. checkpointing), another thread takes the lock in write mode. This prevents any new read operations from taking the lock, effectively stopping any other operations.
- Since the lock isn't asynchronous, those threads are taken and unable to perform any other operations in the mean time.
- Moreover, since the operation is long running and holding a read lock, we can't perform the write until it is over.
- If enough operations are triggered during this time span, we will run out of threads in the ThreadPool.

The intent of this change is to add facilities to cancel ongoing long-running operations when the RocksDb instance is replaced. This is done via a killswitch, which is optionally provided via the Use call, users are expected to guarantee that long running background work will release the lock when the kill switch is triggered, thereby leaving room for the writer to come in and do its job.

We might further attempt to change the lock inside the KeyValueStoreGuard by an async read/write lock, although this is "problematic" because such locks can't be reentrant.

This new mechanism does add some weirdness: it is possible for an ongoing content garbage collection to change databases in the middle of the run. However, this shouldn't affect correctness.