Directory symlink is probed as file
Because directory symlink probe is treated as directory (OpenedDirectory == true), the target path of the directory symlink is not tracked. This can result in underbuild if the target path changes.

The changes also make Detours align with the managed side.

This feature is put behind a flag to not break existing customers.