Merged PR 540288: Split integration test up in smaller chunks to not timeout in the rolling
Split integration test up in smaller chunks to not timeout in the rolling public build.
The public build runs on xunit, not qtest so it supports splitting by category.
We have seen frequent timeouts of this test in the public build part of our rolling build and in the pr validation.