Merged PR 538167: Display CaSaaS offline time
Goal of this work item is to determine the amount of time CaSaaS was "offline" and not receiving operations, then display the offline time in when the service is started and able to receive operations.

Each machine will log periodically to a local file i.e. "CaSaaSRunning.txt" with a message containing a timestamp. When we start up CaSaaS we read the file for the last logged timestamp, which would mark the time CaSaaS stopped receiving operations, and use that timestamp to determine offline time.

Edge cases:
- First time running these bits, file is not created. We create the file, but can not determine previous offline time.
- File is empty (should not happen), can not determine previous offline time

Flag: "ServiceRunningLogInSeconds" in DistributedContentSettings, controls how often we log to the file. If the time interval is too big, the offlineTime will not be as accurate, but if it's too small we will log very often.

KustoLog: Look for messages with  "CaSaaS started", will display offlineTime along with it