Merged PR 540446: Trusted accesses should trump DFAs based on file existence regardless of temporal order
File-existence-based DFAs are applied when a file that is not created by a given pip is written under a shared opaque. Trusted accesses (aka augmented file accesses) do not follow this logic and they are always taken for good (first, because they are trusted, and second, because the 'first' access may not be captured when processes breakaway).
This PR extends this logic to make trusted accesses trump file-existence-based DFAs regardless of the order of arrival. Consider the following scenario:
1) A file is written under a breakaway process (and it is not observed) in the scope of pip A
2) The file is written again in the scope of pip A, but now under a detoured process. A file-existence-based denial is issued since we missed the first access
3) The missing access in 1) is compensated by some means in the scope of pip A.

We want 3) to suppress the DFA, even if happening later than 2). This is a real scenario for MSBuild shared compilation, where the compensating phase happens at the end of the corresponding csc/vbc task.